// src/app/models/persona-request.model.ts
export class PersonaRequest {
    id: number = 0;
    idPersonTypeDocument: number = 0;
    idPersonType: number = 0;
    document: string | null = "";
    name: string | null = "";
    lastNameFirst: string | null = "";
    lastNameSecond: string | null = "";
    bloodType: string | null = "";
    birthDate: string = "";
    idGender: number = 0;
    idStatus: number = 0;
    fullName: string | null = "";
}